// frontend/src/api/model/adminMovieDTO.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Movie Review System API
 * Backend API for Movie Review Website
 * OpenAPI spec version: v1.0
 */
import type { AdminMovieDTOActorsItem } from "./adminMovieDTOActorsItem";
import type { AdminMovieDTODirectorsItem } from "./adminMovieDTODirectorsItem";
import type { AdminMovieDTOWritersItem } from "./adminMovieDTOWritersItem";

/**
 * 后台新增/更新电影请求参数
 */
export interface AdminMovieDTO {
  /**
   * 电影名称
   * @minLength 0
   * @maxLength 200
   */
  name: string;
  /**
   * 电影别名
   * @minLength 0
   * @maxLength 500
   */
  alias?: string;
  /** 演员列表 */
  actors?: AdminMovieDTOActorsItem[];
  /**
   * 封面图片路径
   * @minLength 0
   * @maxLength 500
   */
  cover?: string;
  /** 导演列表 */
  directors?: AdminMovieDTODirectorsItem[];
  /**
   * 本站评分
   * @minimum 0
   * @maximum 10
   */
  score?: number;
  /**
   * 豆瓣评分
   * @minimum 0
   * @maximum 10
   */
  doubanScore?: number;
  /**
   * 本站评分人数
   * @minimum 0
   */
  votes?: number;
  /**
   * 豆瓣评分人数
   * @minimum 0
   */
  doubanVotes?: number;
  /**
   * 电影类型
   * @minLength 0
   * @maxLength 300
   */
  genres?: string;
  /**
   * IMDB ID
   * @minLength 0
   * @maxLength 50
   */
  imdbId?: string;
  /**
   * 语言
   * @minLength 0
   * @maxLength 100
   */
  languages?: string;
  /**
   * 片长
   * @minLength 0
   * @maxLength 50
   */
  mins?: string;
  /**
   * 制片国家/地区
   * @minLength 0
   * @maxLength 200
   */
  regions?: string;
  /**
   * 上映日期
   * @minLength 0
   * @maxLength 100
   */
  releaseDate?: string;
  /**
   * 剧情简介
   * @minLength 0
   * @maxLength 20000
   */
  storyline?: string;
  /**
   * 上映年份
   * @minimum 1888
   * @maximum 2100
   */
  year?: number;
  /** 编剧列表 */
  writers?: AdminMovieDTOWritersItem[];
  /**
   * 上榜理由
   * @minLength 0
   * @maxLength 500
   */
  reason?: string;
}