// frontend/src/api/model/movie.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Movie Review System API
 * Backend API for Movie Review Website
 * OpenAPI spec version: v1.0
 */
import type { MovieActorsItem } from "./movieActorsItem";
import type { MovieDirectorsItem } from "./movieDirectorsItem";
import type { MovieWritersItem } from "./movieWritersItem";

/**
 * 电影详细信息实体
 */
export interface Movie {
  /** 电影ID */
  id?: number;
  /** 电影名称 */
  name?: string;
  /** 电影别名 */
  alias?: string;
  /** 演员列表 (包含name和role等信息) */
  actors?: MovieActorsItem[];
  /** 封面图片路径 (自动拼接域名) */
  cover?: string;
  /** 导演列表 */
  directors?: MovieDirectorsItem[];
  /** 本站评分 (1-10) */
  score?: number;
  /** 豆瓣评分 (1-10) */
  doubanScore?: number;
  /** 本站评分人数 */
  votes?: number;
  /** 豆瓣评分人数 */
  doubanVotes?: number;
  /** 电影类型 (逗号分隔) */
  genres?: string;
  /** IMDB ID */
  imdbId?: string;
  /** 语言 */
  languages?: string;
  /** 片长 */
  mins?: string;
  /** 制片国家/地区 */
  regions?: string;
  /** 上映日期 */
  releaseDate?: string;
  /** 剧情简介 */
  storyline?: string;
  /** 上映年份 */
  year?: number;
  /** 编剧列表 */
  writers?: MovieWritersItem[];
  /** 上榜理由(冷门佳作榜特有) */
  reason?: string;
}